Adapter관련 클래스 분리
